Extract additional checkout field type behavior into per-type classes
Moves the per-type branching in CheckoutFields into a field type class per
supported type, moves value storage into a CheckoutFieldsStorage trait, and
moves the core field definitions into CoreCheckoutFields.
