+++ plugins/woocommerce/src/Blocks/Domain/Services/CheckoutFieldTypes/TextFieldType.php
@@ -1,0 +1,12 @@
+<?php
+declare( strict_types = 1);
+
+namespace Automattic\WooCommerce\Blocks\Domain\Services\CheckoutFieldTypes;
+
+/**
+ * The "text" additional checkout field type.
+ *
+ * Text fields have no type-specific options or constraints, so the AbstractFieldType defaults apply.
+ */
+class TextFieldType extends AbstractFieldType {
+}
